backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
boto3==1.34.32
pydantic==2.5.3
pytube==15.0.0
yt-dlp==2025.03.27
openai==1.14.0
moviepy==1.0.3
ffmpeg-python==0.2.0
python-dotenv==1.0.0
mangum==0.17.0
